toolkit/components/extensions/test/mochitest/test_ext_notifications.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for notifications</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="text/javascript">
"use strict";

add_task(function* test_notifications() {
  function backgroundScript() {
    browser.test.log("running background script");

    let opts = {title: "Testing Notification", message: "Carry on"};

    // Test an unimplemented listener.
    browser.notifications.onClicked.addListener(function() {});

    browser.notifications.create("5", opts, function(id) {
      browser.test.sendMessage("running", id);
      browser.test.notifyPass("background test passed");
    });
  }

  let extensionData = {
    manifest: {
      permissions: ["notifications"],
    },
    background: "(" + backgroundScript.toString() + ")()",
  };

  let extension = ExtensionTestUtils.loadExtension(extensionData);
  info("load complete");
  yield extension.startup();
  info("startup complete");
  let x = yield extension.awaitMessage("running");
  is(x, "5", "got correct value from extension");
  yield extension.awaitFinish();
  info("test complete");
  yield extension.unload();
  info("extension unloaded successfully");
});

add_task(function* test_notifications_empty_getAll() {
  function backgroundScript() {
    browser.test.log("running background script");

    browser.notifications.getAll(notifications => {
      browser.test.assertTrue(Array.isArray(notifications),
        "getAll() returned an array");
      browser.test.assertEq(notifications.length, 0, "the array was empty");
      browser.test.notifyPass("getAll empty");
    });
  }

  let extensionData = {
    manifest: {
      permissions: ["notifications"],
    },
    background: "(" + backgroundScript.toString() + ")()",
  };

  let extension = ExtensionTestUtils.loadExtension(extensionData);
  info("load complete");
  yield extension.startup();
  info("startup complete");
  yield extension.awaitFinish("getAll empty");
  info("test complete");
  yield extension.unload();
  info("extension unloaded successfully");
});

add_task(function* test_notifications_populated_getAll() {
  function backgroundScript() {
    browser.test.log("running background script");

    let opts = {title: "Testing Notification", message: "Carry on"};
    browser.notifications.create("p1", opts, () => {
      browser.notifications.create("p2", opts, () => {
        browser.notifications.getAll(notifications => {
          browser.test.assertTrue(Array.isArray(notifications),
            "getAll() returned an array");
          browser.test.assertEq(notifications.length, 2,
            "the array contained two notification ids");
          browser.test.assertTrue(notifications.includes("p1"),
            "the array contains the first notification");
          browser.test.assertTrue(notifications.includes("p2"),
            "the array contains the second notification");
          browser.test.notifyPass("getAll populated");
        });
      });
    });
  }

  let extensionData = {
    manifest: {
      permissions: ["notifications"],
    },
    background: "(" + backgroundScript.toString() + ")()",
  };

  let extension = ExtensionTestUtils.loadExtension(extensionData);
  info("load complete");
  yield extension.startup();
  info("startup complete");
  yield extension.awaitFinish("getAll populated");
  info("test complete");
  yield extension.unload();
  info("extension unloaded successfully");
});

</script>

</body>
</html>
